# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AxionID API",
        "description": "Documentação oficial do sistema de identidade AxionID",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/v1/forgot-password": {
            "post": {
                "tags": [
                    "Recuperação de Senha"
                ],
                "summary": "1. Solicitar código de recuperação",
                "operationId": "727cb7858ffeb4e102be8ab66681e1b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "josilva1233@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código enviado"
                    }
                }
            }
        },
        "/api/v1/verify-code": {
            "post": {
                "tags": [
                    "Recuperação de Senha"
                ],
                "summary": "2. Validar código recebido",
                "operationId": "fffb18746ad52ddb67dd0ef49a72b999",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "josilva1233@gmail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "ABC123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código válido"
                    }
                }
            }
        },
        "/api/v1/reset-password": {
            "post": {
                "tags": [
                    "Recuperação de Senha"
                ],
                "summary": "3. Definir nova senha",
                "operationId": "10071ce688d34466c60a6ba60a654cf4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "josilva1233@gmail.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "ABC123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nova_senha123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "nova_senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token gerado no login (ex: 1|abcde...) para autorizar as chamadas.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}